+++ R/generatePlotData.R
@@ -27,46 +27,6 @@ CovCalc <- function(bamFile, windowSize=250){
   return(cov)
 }
 
-checkIREquality <- function(gbkData, regions){
-  
-  gbkSeq <- genbankr::getSeq(gbkData)
-
-  repeatB <- as.numeric(regions[which(regions[,4] == "IRb"),2:3])
-  repeatA <- as.numeric(regions[which(regions[,4] == "IRa"),2:3])
-  if(repeatB[2] - repeatB[1] != repeatA[2] - repeatA[1]){
-    message("WARNING: Inverted repeats differ in sequence length")
-    message(paste("The IRb has a total lengths of: ", repeatB[2]-repeatB[1], " bp", sep=""))
-    message(paste("The IRa has a total lengths of: ", repeatA[2]-repeatA[1], " bp", sep=""))
-  }
-  if(gbkSeq[[1]][repeatB[1]:repeatB[2]] != Biostrings::reverseComplement(gbkSeq[[1]][repeatA[1]:repeatA[2]])){
-    IRa_seq <- Biostrings::DNAString(gbkSeq[[1]][repeatB[1]:repeatB[2]])
-    IRa_seq <- split(IRa_seq, ceiling(seq_along(IRa_seq)/10000))
-    IRb_seq <- Biostrings::DNAString(Biostrings::reverseComplement(gbkSeq[[1]][repeatA[1]:repeatA[2]]))
-    IRb_seq <- split(IRb_seq, ceiling(seq_along(IRb_seq)/10000))
-    
-    IR_diff_SNPS <- c()
-    IR_diff_gaps <- c()
-    for(i in  1:min(length(IRa_seq), length(IRb_seq))){
-      subst_mat <- Biostrings::nucleotideSubstitutionMatrix(match=1, mismatch=-3, baseOnly=TRUE)
-      globalAlign <- tryCatch(
-        {Biostrings::pairwiseAlignment(IRa_seq[[i]], IRb_seq[[i]], substitutionMatrix=subst_mat, gapOpening=5, gapExtension=2)},
-        error = function(e) {return(NULL)})
-      if(is.null(globalAlign)) break
-      IR_diff_SNPS <- c(IR_diff_SNPS, which(strsplit(Biostrings::compareStrings(globalAlign), "")[[1]]=="?"))
-      IR_diff_gaps <- c(IR_diff_gaps, which(strsplit(Biostrings::compareStrings(globalAlign), "")[[1]]=="-"))
-    }
-    message("WARNING: Inverted repeats differ in sequence")
-    if(length(IR_diff_SNPS)>0){
-      message(paste("When aligned, the IRs differ through a total of ", length(IR_diff_SNPS)," SNPS. These SNPS are located at the following nucleotide positions: ", paste(unlist(IR_diff_SNPS), collapse=" "), sep=""))
-    }
-    if(length(IR_diff_gaps)>0){
-      message(paste("When aligned, the IRs differ through a total of ", length(IR_diff_gaps)," gaps. These gaps are located at the following nucleotide positions: ", paste(unlist(IR_diff_gaps), collapse=" "), sep=""))
-    }
-  }
-  message("Proceeding with coverage depth visualization, but without quadripartite genome structure ...")
-}
-
-
 GenerateIRSynteny <- function(genes, syntenyLineType) {
   
   n_occur <- data.frame(table(genes[,4]),stringsAsFactors = FALSE)